+++ geopandas/array.py
@@ -333,17 +333,6 @@ class GeometryArray(ExtensionArray):
         self._sindex = None
 
     @property
-    def data(self):
-        warnings.warn(
-            "Accessing the underlying geometries through the `.data` attribute is "
-            "deprecated and will be removed in GeoPandas 1.0. You can use "
-            "`np.asarray(..)` or the `to_numpy()` method instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return self._data
-
-    @property
     def sindex(self):
         if self._sindex is None:
             self._sindex = SpatialIndex(self._data)